net/portmapper: add upnp port mapping
This is a subset of my original PR (#2113) which had too many design changes.
I've distilled most of it into just the changes necessary for portmapping.
This also removes the upnp probing, which seemed to be broken. A future change can add it
back in. but this makes upnp mapping work without probing.

Signed-off-by: julianknodt <julianknodt@gmail.com>